Moved helpers to their own files
