+++ src/main/java/frc/robot/autonomous/Chooser.java
@@ -1,6 +1,5 @@
 package frc.robot.autonomous;
 
-import java.util.List;
 
 import edu.wpi.first.wpilibj.geometry.Pose2d;
 import edu.wpi.first.wpilibj.geometry.Rotation2d;
@@ -16,6 +15,9 @@ import edu.wpi.first.wpilibj2.command.WaitUntilCommand;
 import frc.lib5k.utils.RobotLogger;
 import frc.robot.RobotConstants;
 import frc.robot.autonomous.actions.LogCommand;
+import frc.robot.autonomous.helpers.EasyTrajectory;
+import frc.robot.autonomous.helpers.PathGenerator;
+import frc.robot.autonomous.helpers.SpeedConstraint;
 import frc.robot.subsystems.DriveTrain;
 
 /**
@@ -91,9 +93,9 @@ public class Chooser {
 
         // Test path following
         SequentialCommandGroup path = PathGenerator.generate(
-                new PathGenerator.EasyTrajectory(new Pose2d(0, 0, new Rotation2d(0)),
+                new EasyTrajectory(new Pose2d(0, 0, new Rotation2d(0)),
                         new Pose2d(3, 0, Rotation2d.fromDegrees(0))),
-                new PathGenerator.SpeedConstraint(1.0, 1.0));
+                new SpeedConstraint(1.0, 1.0));
 
         outputCommand.addCommands(path);
         // /* Start building command based on params */
