+++ src/main/java/frc/robot/autonomous/helpers/SpeedConstraint.java
@@ -1,0 +1,12 @@
+package frc.robot.autonomous.helpers;
+
+public class SpeedConstraint {
+    public double maxSpeedPercent;
+    public double maxAccelPercent;
+
+    public SpeedConstraint(double maxSpeedPercent, double maxAccelPercent) {
+        this.maxAccelPercent = maxAccelPercent;
+        this.maxSpeedPercent = maxSpeedPercent;
+
+    }
+}
